+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "github-diff-enhancer",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "keywords": [
     "Chrome Extension",
     "GitHub"
